Split HeadTracker into its own file
